requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Transformers ecosystem
transformers>=4.38.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0
tokenizers>=0.15.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.1
evaluate>=0.4.1

# Experiment tracking (optional — remove if not needed)
wandb>=0.16.0
mlflow>=2.11.0

# AWS
boto3>=1.34.0

# Config / utilities
pyyaml>=6.0
tqdm>=4.66.0

# Bias/toxicity model dependencies are handled by transformers.
# Optional: bitsandbytes for 4-bit quantized training (Linux only)
# bitsandbytes>=0.43.0

# Type checking / linting (dev only)
# mypy>=1.9.0
# ruff>=0.3.0
